Rename ws package to http
