+++ cmd/create-channels/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/statechannels/go-nitro/internal/logging"
 	"github.com/statechannels/go-nitro/internal/testdata"
 	"github.com/statechannels/go-nitro/rpc"
-	"github.com/statechannels/go-nitro/rpc/transport/ws"
+	"github.com/statechannels/go-nitro/rpc/transport/http"
 	"github.com/statechannels/go-nitro/types"
 )
 
@@ -40,7 +40,7 @@ func createChannels() error {
 		logging.SetupDefaultFileLogger(LOG_FILE, slog.LevelDebug)
 
 		url := fmt.Sprintf(":%d/api/v1", participantOpts.RpcPort)
-		clientConnection, err := ws.NewWebSocketTransportAsClient(url)
+		clientConnection, err := http.NewWebSocketTransportAsClient(url)
 		if err != nil {
 			return err
 		}
