+++ internal/rpc/rpc.go
@@ -7,8 +7,8 @@ import (
 	"github.com/statechannels/go-nitro/node"
 	"github.com/statechannels/go-nitro/rpc"
 	"github.com/statechannels/go-nitro/rpc/transport"
+	"github.com/statechannels/go-nitro/rpc/transport/http"
 	"github.com/statechannels/go-nitro/rpc/transport/nats"
-	"github.com/statechannels/go-nitro/rpc/transport/ws"
 )
 
 func InitializeRpcServer(node *node.Node, rpcPort int, useNats bool) (*rpc.RpcServer, error) {
@@ -20,7 +20,7 @@ func InitializeRpcServer(node *node.Node, rpcPort int, useNats bool) (*rpc.RpcSe
 		transport, err = nats.NewNatsTransportAsServer(rpcPort)
 	} else {
 		slog.Info("Initializing websocket RPC transport...")
-		transport, err = ws.NewHttpSocketTransportAsServer(fmt.Sprint(rpcPort))
+		transport, err = http.NewHttpSocketTransportAsServer(fmt.Sprint(rpcPort))
 	}
 	if err != nil {
 		return nil, err
