+++ rpc/transport/http/client.go
@@ -1,4 +1,4 @@
-package ws
+package http
 
 import (
 	"bytes"
